Remove some logging and add sell functionality
