+++ src/components/marketplace-components/Filters.js
@@ -76,7 +76,6 @@ function Filters() {
     };
     const handleSubmit = e => {
         e.preventDefault()
-        console.log(values)
     }
     const {
         values,
